# backend/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "dataset": {
            "default": "operational",
            "enum": [
              "operational",
              "sample"
            ],
            "title": "Dataset",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post",
        "type": "object"
      },
...
      "DocumentDetail": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "datasetReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datasetreason"
          },
          "datasetScope": {
            "default": "operational",
            "enum": [
              "operational",
              "sample"
            ],
            "title": "Datasetscope",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentError"
              },
              {
                "type": "null"
              }
            ]
          },
          "extraction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction"
          },
          "fields": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Fields",
            "type": "array"
          },
          "fileName": {
            "title": "Filename",
            "type": "string"
          },
          "fileSize": {
            "minimum": 0.0,
            "title": "Filesize",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isSynthetic": {
            "default": false,
            "title": "Issynthetic",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "mimeType": {
            "title": "Mimetype",
            "type": "string"
          },
          "ocrPages": {
            "items": {
              "$ref": "#/components/schemas/OCRPage"
            },
            "title": "Ocrpages",
            "type": "array"
          },
          "pageCount": {
            "minimum": 1.0,
            "title": "Pagecount",
            "type": "integer"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/DocumentPage"
            },
            "title": "Pages",
            "type": "array"
          },
          "processedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processedat"
          },
          "processingMetadata": {
            "additionalProperties": true,
            "title": "Processingmetadata",
            "type": "object"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "reviewStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoreBreakdown": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scorebreakdown"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageProgress"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "title": "Uploadedat",
            "type": "string"
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "fileName",
          "mimeType",
          "fileSize",
          "pageCount",
          "uploadedAt",
          "status",
          "stage",
          "updatedAt"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
      "DocumentError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "DocumentError",
        "type": "object"
      },
      "DocumentList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "DocumentList",
        "type": "object"
      },
...
      "DocumentSummary": {
        "additionalProperties": false,
        "properties": {
          "datasetReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datasetreason"
          },
          "datasetScope": {
            "default": "operational",
            "enum": [
              "operational",
              "sample"
            ],
            "title": "Datasetscope",
            "type": "string"
          },
          "fileName": {
            "title": "Filename",
            "type": "string"
          },
          "fileSize": {
            "minimum": 0.0,
            "title": "Filesize",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isSynthetic": {
            "default": false,
            "title": "Issynthetic",
            "type": "boolean"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "mimeType": {
            "title": "Mimetype",
            "type": "string"
          },
          "pageCount": {
            "minimum": 1.0,
            "title": "Pagecount",
            "type": "integer"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "reviewStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "uploadedAt": {
            "format": "date-time",
            "title": "Uploadedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fileName",
          "mimeType",
          "fileSize",
          "pageCount",
          "uploadedAt",
          "status",
          "stage"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OCRPage": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "pageNumber": {
            "minimum": 1.0,
            "title": "Pagenumber",
            "type": "integer"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/OCRRegion"
            },
            "title": "Regions",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "pageNumber",
          "text",
          "width",
          "height"
        ],
        "title": "OCRPage",
        "type": "object"
      },
...
      "ProcessingStatus": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageProgress"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "stage",
          "attempts",
          "stages",
          "updatedAt"
        ],
        "title": "ProcessingStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-assisted land-record digitization prototype. Validation checks support human verification and do not determine legal ownership.",
    "title": "BhuDrishti AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "dataset",
            "required": false,
            "schema": {
              "default": "operational",
              "enum": [
                "operational",
                "sample"
              ],
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List uploaded documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload a real PDF or image",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dataset",
            "required": false,
            "schema": {
              "default": "operational",
              "enum": [
                "operational",
                "sample"
              ],
              "title": "Dataset",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get uploaded document and evidence",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/pages/{page_number}/image": {
      "get": {
        "operationId": "page_image_api_v1_documents__document_id__pages__page_number__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_number",
            "required": true,
            "schema": {
              "title": "Page Number",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "dataset",
            "required": false,
            "schema": {
              "default": "operational",
              "enum": [
                "operational",
                "sample"
              ],
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "original",
            "required": false,
            "schema": {
              "default": false,
              "title": "Original",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "View rendered document page",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/process": {
      "post": {
        "operationId": "process_document_api_v1_documents__document_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dataset",
            "required": false,
            "schema": {
              "default": "operational",
              "enum": [
                "operational",
                "sample"
              ],
              "title": "Dataset",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start or retry real document processing",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/processing-status": {
      "get": {
        "operationId": "processing_status_api_v1_documents__document_id__processing_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dataset",
            "required": false,
            "schema": {
              "default": "operational",
              "enum": [
                "operational",
                "sample"
              ],
              "title": "Dataset",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Poll actual processing stage",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}